Yolov8_train_extraction_and_analysis/requirements.txt
notebook==7.1.3
opencv-python==4.7.0.72
matplotlib==3.7.1
numpy==1.24.3
pandas==1.5.3
plotly==5.21.0
pyyaml==6.0.1
tqdm==4.65.0
python-dotenv==1.0.1
isort
ultralytics==8.2.1
py-markdown-table
onnx==1.16.0
onnx-simplifier==0.4.36
torchinfo==1.8.0
tensorboard==2.13.0
black==22.12.0